* test for xfem corrected \n * removed warnings
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@3209 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ features/features.cpp
@@ -5612,7 +5612,7 @@ bool FeatureTree::stepInternal(bool guided, bool xfemIteration)
         }
     }
 
-    int it = 0 ;
+    size_t it = 0 ;
     bool needexit = false ;
     do
     {
